archive/webapp_v2_backend/requirements.txt
# FastAPI Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.3
asyncpg==0.29.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0

# Scientific Computing (for on-demand similarity)
numpy==1.26.2
numba==0.58.1

# Testing
pytest==8.3.4
pytest-cov==6.0.0
pytest-timeout==2.3.1
httpx==0.27.2  # Required for TestClient
